+++ R/shinychat-package.R
@@ -1,0 +1,9 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom coro async
+#' @importFrom htmltools tag css HTML
+#' @importFrom rlang %||%
+## usethis namespace: end
+NULL