+++ tests/testthat/test-chat.R
@@ -1,3 +1,43 @@
-test_that("UI works", {
+library(htmltools)
+
+test_that("Chat component markup", {
   expect_snapshot(chat_ui("chat"))
+
+  expect_snapshot({
+    chat_ui("chat", messages = list("Foo", "Bar"))
+  })
+
+  expect_snapshot({
+    chat_ui(
+      "chat",
+      messages = list(
+        list(content = "Assistant", role = "assistant"),
+        list(content = "User", role = "user")
+      )
+    )
+  })
+
+  expect_snapshot({
+    chat_ui(
+      "chat",
+      messages = list(
+        div("Hello"),
+        span("world")
+      )
+    )
+  })
+
+  expect_snapshot({
+    render_tags(
+      chat_ui(
+        "chat",
+        messages = list(
+          div("Hello", htmlDependency("foo", "1.0.0", "")),
+          span("world")
+        )
+      )
+    )
+  })
+
+  # TODO: it'd be nice to mock the shinyChatMessage custom messages
 })